+++ micro-ui/web/micro-ui-internals/packages/modules/core/src/pages/employee/Login/login.js
@@ -4,12 +4,13 @@ import React, { useEffect, useState } from "react";
 import { useHistory } from "react-router-dom";
 import Background from "../../../components/Background";
 import Header from "../../../components/Header";
+import ImageComponent from "../../../components/ImageComponent";
 
 /* set employee details to enable backward compatiable */
 const setEmployeeDetail = (userObject, token) => {
- if (Digit.Utils.getMultiRootTenant()) {
-     return;
-   }  
+  if (Digit.Utils.getMultiRootTenant() && process.env.NODE_ENV !== "development") {
+    return;
+  }
   let locale = JSON.parse(sessionStorage.getItem("Digit.locale"))?.value || Digit.Utils.getDefaultLanguage();
   localStorage.setItem("Employee.tenant-id", userObject?.tenantId);
   localStorage.setItem("tenant-id", userObject?.tenantId);
@@ -165,13 +166,12 @@ const Login = ({ config: propsConfig, t, isDisabled, loginOTPBased }) => {
     // Set disable based on the check
     setDisable(hasEmptyFields);
   };
-
   return isLoading || isStoreLoading ? (
     <Loader />
   ) : (
     <Background>
       <div className="employeeBackbuttonAlign">
-      <BackLink onClick={() => window.history.back()}/>
+        <BackLink onClick={() => window.history.back()} />
       </div>
       <FormComposerV2
         onSubmit={loginOTPBased ? onOtpLogin : onLogin}
@@ -190,11 +190,11 @@ const Login = ({ config: propsConfig, t, isDisabled, loginOTPBased }) => {
         cardClassName="loginCardClassName"
         buttonClassName="buttonClassName"
       >
-        {stateInfo?.code ? <Header /> : <Header showTenant={false} /> }
+        {stateInfo?.code ? <Header /> : <Header showTenant={false} />}
       </FormComposerV2>
       {showToast && <Toast type={"error"} label={t(showToast)} onClose={closeToast} />}
       <div className="employee-login-home-footer" style={{ backgroundColor: "unset" }}>
-        <img
+        <ImageComponent
           alt="Powered by DIGIT"
           src={window?.globalConfigs?.getConfig?.("DIGIT_FOOTER_BW")}
           style={{ cursor: "pointer" }}