+++ micro-ui/web/micro-ui-internals/packages/modules/core/src/pages/employee/Otp/index.js
@@ -1,10 +1,11 @@
-import React, { useState ,useEffect} from "react";
+import React, { useState, useEffect } from "react";
 import { BackLink, CardLabel, Loader, Toast } from "@egovernments/digit-ui-components";
 import { FormComposerV2 } from "@egovernments/digit-ui-react-components";
 import { useTranslation } from "react-i18next";
 import { Route, Switch, useRouteMatch, useHistory, useLocation } from "react-router-dom";
 import Background from "../../../components/Background";
 import Header from "../../../components/Header";
+import ImageComponent from "../../../components/ImageComponent";
 
 /* set employee details to enable backward compatiable */
 const setEmployeeDetail = (userObject, token) => {
@@ -32,7 +33,7 @@ const Otp = ({ isLogin = false }) => {
   const [isOtpValid, setIsOtpValid] = useState(false);
   const [user, setUser] = useState(null);
   const [params, setParams] = useState(location?.state?.data || {});
-  const [ifSuperUserExists, setIfSuperUserExist]= useState(false);
+  const [ifSuperUserExists, setIfSuperUserExist] = useState(false);
   const { email, tenant } = location.state || {};
   const { data: MdmsRes } = Digit.Hooks.useCustomMDMS(
     tenant,
@@ -44,16 +45,16 @@ const Otp = ({ isLogin = false }) => {
     ],
     {
       enabled: true,
-      staleTime:0,
-      cacheTime:0,
+      staleTime: 0,
+      cacheTime: 0,
       select: (data) => {
         return data?.["SandBoxLanding"]?.["LandingPageRoles"];
       },
     }
   );
 
-  const RoleLandingUrl= MdmsRes?.[0].url;
-  
+  const RoleLandingUrl = MdmsRes?.[0].url;
+
   const config = [
     {
       body: [
@@ -100,27 +101,25 @@ const Otp = ({ isLogin = false }) => {
     setEmployeeDetail(user?.info, user?.access_token);
     let redirectPath = `/${window?.globalPath}/user/setup`;
 
-
     const getRedirectPathOtpLogin = (locationPathname, user, MdmsRes, RoleLandingUrl) => {
       const userRole = user?.info?.roles?.[0]?.code;
       const isSuperUser = userRole === "SUPERUSER";
       const contextPath = window?.contextPath;
-  
-      switch (true) {
-          case locationPathname === "/sandbox-ui/user/otp" && isSuperUser:
-              return `/${contextPath}/employee/user/landing`;
-  
-          case isSuperUser && MdmsRes?.[0]?.rolesForLandingPage?.includes("SUPERUSER"):
-              return `/${contextPath}${RoleLandingUrl}`;
-  
-          default:
-              return `/${contextPath}/employee`;
-      }
-  };
-  
-  // Usage
-  const redirectPathOtpLogin = getRedirectPathOtpLogin(location.pathname, user, MdmsRes, RoleLandingUrl);
 
+      switch (true) {
+        case locationPathname === "/sandbox-ui/user/otp" && isSuperUser:
+          return `/${contextPath}/employee/user/landing`;
+
+        case isSuperUser && MdmsRes?.[0]?.rolesForLandingPage?.includes("SUPERUSER"):
+          return `/${contextPath}${RoleLandingUrl}`;
+
+        default:
+          return `/${contextPath}/employee`;
+      }
+    };
+
+    // Usage
+    const redirectPathOtpLogin = getRedirectPathOtpLogin(location.pathname, user, MdmsRes, RoleLandingUrl);
 
     if (isLogin) {
       history.push(redirectPathOtpLogin);
@@ -185,7 +184,7 @@ const Otp = ({ isLogin = false }) => {
       </FormComposerV2>
       {showToast && <Toast type={"error"} label={t(showToast)} onClose={closeToast} />}
       <div className="EmployeeLoginFooter">
-        <img
+        <ImageComponent
           alt="Powered by DIGIT"
           src={window?.globalConfigs?.getConfig?.("DIGIT_FOOTER_BW")}
           style={{ cursor: "pointer" }}