+++ micro-ui/web/micro-ui-internals/packages/modules/core/src/pages/employee/index.js
@@ -16,6 +16,7 @@ import UserProfile from "../citizen/Home/UserProfile";
 import ErrorComponent from "../../components/ErrorComponent";
 import { PrivateRoute } from "@egovernments/digit-ui-components";
 import RoleLanding from "./RoleLanding";
+import ImageComponent from "../../components/ImageComponent";
 
 const userScreensExempted = ["user/landing", "user/profile", "user/error"];
 
@@ -145,7 +146,7 @@ const EmployeeApp = ({
               </ErrorBoundary>
             </div>
             <div className="employee-home-footer">
-              <img
+              <ImageComponent
                 alt="Powered by DIGIT"
                 src={window?.globalConfigs?.getConfig?.("DIGIT_FOOTER")}
                 style={{ height: "1.1em", cursor: "pointer" }}