+++ micro-ui/web/micro-ui-internals/packages/modules/sandbox/src/pages/employee/tenantMgmt/LogoUploaderComponent.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import { CardLabel, LabelFieldPair, Uploader } from "@egovernments/digit-ui-components";
+import { CardLabel, LabelFieldPair, FileUpload } from "@egovernments/digit-ui-components";
 import {Toast} from "@egovernments/digit-ui-react-components";
 import { useTranslation } from "react-i18next";
 import { Fragment } from "react";
@@ -52,9 +52,9 @@ const LogoUploaderComponent = ({ onSelect, ...props }) => {
     <>
       <LabelFieldPair className={"uploader-label-field"}>
         <CardLabel>{`${t("LOGO_UPLOAD_LABEL")}`}</CardLabel>
-        <Uploader
+        <FileUpload
           uploadedFiles={[]}
-          variant="uploadFile"
+          variant="uploadField"
           onUpload={(files) => selectFile(files)}
           iserror={uploadErrorMessage}
           accept=".jpg, .png, .jpeg"